+++ packages/benchmarks/css-tokenizer-benchmark/src/utils/tables.ts
@@ -1,8 +1,9 @@
 /**
  * @file Write benchmark results to a markdown file
  */
-import { formatRFC7231 } from 'date-fns';
 import { writeFile } from 'fs/promises';
+
+import { formatRFC7231 } from 'date-fns';
 import { markdownTable } from 'markdown-table';
 import si from 'systeminformation';